modify code for reset logic
