+++ LivingDesign/View/Delete.swift
@@ -15,6 +15,7 @@ class Delete: UIViewController, UICollectionViewDataSource, UICollectionViewDele
     @IBOutlet weak var collectionView: UICollectionView!
     
     private let wireframe: RootViewWireframe = RootViewWireframe()
+    private let resetController: ResetController = ResetController()
     
     let realm = try! Realm()
     
@@ -76,9 +77,10 @@ class Delete: UIViewController, UICollectionViewDataSource, UICollectionViewDele
         for dci in deleteCandidateItems{
             try! realm.write {
                 realm.delete(realm.objects(Item.self).filter("id == %@",dci.id).first!)
-                realm.delete(dci)
             }
         }
+        
+        self.resetController.reset()
         
         let nextStoryBoard = UIStoryboard(name: "Home", bundle: nil)
         let nextViewController = nextStoryBoard.instantiateViewController(withIdentifier: "HomeViewControllerID")
@@ -87,15 +89,7 @@ class Delete: UIViewController, UICollectionViewDataSource, UICollectionViewDele
     
     // 取り消しボタンを押すと、削除リストを削除して、削除画面に戻す（背景色を戻したい）
     @IBAction func touchUpInsideCancelButton(_ sender: Any) {
-        let deleteCandidateItems = realm.objects(DeleteCandidateItem.self)
-
-        if(deleteCandidateItems.first != nil){
-            for dci in deleteCandidateItems{
-                try! realm.write {
-                    realm.delete(dci)
-                }
-            }
-        }
+        self.resetController.reset()
         
         let nextStoryBoard = UIStoryboard(name: "Delete", bundle: nil)
         let nextViewController = nextStoryBoard.instantiateViewController(withIdentifier: "DeleteViewControllerID")
